Fix incorrect start and stop indices on - strand
Applies a quick fix which simply returns the value of the other
function.  For example, calling get_start_indices on the "-" strand
utilizes the code from get_stop_indices on the "+" strand and vice
versa.

This works because the sequence is already reverse complemented and the
first base in the sequence string is actually the last
position/coordinate and the last base in the sequence string is actually
the first position/coordinate. By utilizing code from the opposite
function, we get the relevant coordinates.

The old behavior would give us the position/coordinates of the end of
the last CDS(lowest position) as the start, and the position/coordinates of the
beginning of the first CDS(highest position) as the stop. The result is
backwards as it assumes the sequence string is in its original order
where the last base position is the start of the sequence. It also grabs
the wrong index from the list of start and stop indices, which causes
the positions to reference the last inner position of the first and last
CDSes.

Fixes: #166